--- components/Tutorial/Step10.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="tutorial-adaptive">
-    <div class="wrapStep1">
-      <h1>Step10</h1>
-    </div>
-  </div>
-</template>
-
-<script>
-export default {}
-</script>
-
-<style lang="scss" scoped></style>
