+++ pages/tutorial/step10.vue
@@ -6,113 +6,56 @@
 
     <div class="wrapStep10">
       <div class="hostel">
-        <div class="itemDesc">
-          Напишите название вашего объекта
+        <div class="title">
+          Итак, нам осталось только открыть ваш объект для бронирования
         </div>
-        <div class="hostelInfoInput">
+        <div class="desc">
+          Для этого установите статус вашего жилья с помощью переключателя ниже.
+          Если ваш объект уже свободен и готов принимать гостей, установите
+          статус «Свободен», если освобождается завтра или послезавтра, то
+          «Освобождается завтра» или «Освобождается послезавтра». Если ваше
+          жильё не освобождается в ближайшие два дня или вы передумали его
+          сдавать, установите статус «Занят».
+        </div>
+        <div class="statusInput">
           <b-form-group
-            id="objectNameLabel"
+            id="objectStatusLabel"
             class="form-group-restork"
-            label="Название объекта"
+            label="Статус объекта"
             label-for="objectNameInput"
           >
             <b-form-input
-              id="objectNameInput"
-              placeholder="Название"
+              id="objectStatusInput"
+              placeholder="Статус объекта"
               type="text"
             ></b-form-input>
           </b-form-group>
         </div>
-
-        <div class="itemDesc secondDesc">
-          Расскажите, чем интересен ваш район, и укажите особенности жилья,
-          например быстрый Wi-Fi или парковку
+        <div class="hintInsideMain">
+          <HintStep11 />
         </div>
-
-        <b-form-group
-          id="objectNameLabel"
-          class="form-group-restork"
-          label="Название объекта"
-          label-for="objectNameInput"
-        >
-          <b-form-textarea
-            id="featuresTextarea"
-            placeholder="Описание"
-            rows="3"
-            max-rows="10"
-          ></b-form-textarea>
-        </b-form-group>
       </div>
     </div>
     <div class="wrapHint">
-      <HintStep1
-        :list="[
-          {
-            name: 'Single',
-            desc:
-              'Однокомнатный, для размещения одного отдыхающего. Спальное место одно.'
-          },
-          {
-            name: 'Double',
-            desc:
-              'Однокомнатный номер для двоих. Может быть с одной большой кроватью или с двумя раздельными кроватями.'
-          },
-          {
-            name: 'Triple',
-            desc: 'Номер для троих отдыхающих.'
-          },
-          {
-            name: 'Family Room',
-            desc: 'Семейный двух- или трехкомнатный номер.'
-          }
-        ]"
-      />
+      <HintStep11 />
     </div>
     <div class="wrapProgress">
       <ProgressBar />
     </div>
-    <PopupStep />
   </div>
 </template>
 
 <script>
 import ProgressBar from '~/components/Tutorial/ProgressBar'
 import Title from '~/components/Tutorial/Title'
-import PopupStep from '~/components/Tutorial/PopupStep'
-import HintStep1 from '~/components/Tutorial/HintStep1'
+import HintStep11 from '~/components/Tutorial/HintStep11'
 
 export default {
-  data() {
-    return {
-      currentStepNumber: 1,
-      iObjectTypeID: 0,
-      iCategoryID: 0,
-      roomCategory: [
-        { id: 1, title: 'Single' },
-        { id: 2, title: 'Double' },
-        { id: 3, title: 'Triple' },
-        { id: 4, title: 'Family room' }
-      ]
-    }
-  },
   layout: 'dashboardV2Tutorial',
   components: {
     Title,
     ProgressBar,
-    PopupStep,
-    HintStep1
-  },
-  computed: {
-    objectType() {
-      return this.$store.state.objectType.list
-    },
-    object() {
-      return this.$store.state.tutorial.object
-    }
-  },
-  methods: {},
-  async asyncData({ store, params }) {
-    await store.dispatch('objectType/GET_LIST')
+    HintStep11
   }
 }
 </script>
@@ -124,32 +67,60 @@ export default {
   grid-template-rows: auto 1fr auto;
   height: calc(100vh - 63px);
   @media (max-width: 991px) {
+    grid-template-columns: 1fr 1fr;
+  }
+  @media (max-width: 767px) {
     grid-template-columns: 1fr;
   }
   .wrapTitle {
     // background: blue;
     grid-column: 1/2;
     grid-row: 1/2;
+    @media (max-width: 991px) {
+      grid-column: 1/3;
+    }
+    @media (max-width: 767px) {
+      grid-column: 1/2;
+    }
   }
   .wrapStep10 {
     // background: red;
     grid-column: 1/2;
     grid-row: 2/3;
     overflow: auto;
-    .itemDesc {
-      margin: 0 0 15px;
-      font-size: 20px;
-      line-height: 26px;
+    margin: -15px;
+    padding: 15px;
+    .title {
+      margin: 0 0 25px;
+      font-size: 30px;
+      line-height: 36px;
       font-weight: 500;
-      @media (max-width: 767px) {
-        font-size: 16px;
-        line-height: 20px;
+      @media (max-width: 991px) {
+        margin: 0 0 15px;
       }
-      &.secondDesc {
-        margin-top: 45px;
-        @media (max-width: 767px) {
-          margin-top: 37px;
-        }
+      @media (max-width: 767px) {
+        font-size: 26px;
+        line-height: 28px;
+        margin: 0 0 20px;
+      }
+    }
+    .desc {
+      margin: 0 0 25px;
+      font-size: 16px;
+      line-height: 26px;
+      font-weight: 400;
+      @media (max-width: 991px) {
+        margin: 0 0 20px;
+      }
+      @media (max-width: 767px) {
+        font-size: 14px;
+        line-height: 20px;
+        margin: 0 0 30px;
+      }
+    }
+    .hintInsideMain {
+      @media (min-width: 768px) {
+        display: none;
       }
     }
   }
@@ -159,8 +130,13 @@ export default {
     grid-row: 2/4;
     overflow: auto;
     display: flex;
-    justify-content: flex-end;
+    justify-content: center;
     @media (max-width: 991px) {
+      //   display: none;
+      grid-column: 2/3;
+      grid-row: 2/3;
+    }
+    @media (max-width: 767px) {
       display: none;
     }
   }
@@ -168,6 +144,12 @@ export default {
     // background: yellow;
     grid-column: 1/2;
     grid-row: 3/4;
+    @media (max-width: 991px) {
+      grid-column: 1/3;
+    }
+    @media (max-width: 767px) {
+      grid-column: 1/2;
+    }
   }
 }
 </style>
