+++ pages/tutorial/step2.vue
@@ -330,14 +330,12 @@
     <div class="wrapProgress">
       <ProgressBar />
     </div>
-    <PopupStep />
   </div>
 </template>
 
 <script>
 import ProgressBar from '~/components/Tutorial/ProgressBar'
 import Title from '~/components/Tutorial/Title'
-import PopupStep from '~/components/Tutorial/PopupStep'
 import HintStep1 from '~/components/Tutorial/HintStep1'
 
 export default {
@@ -358,7 +356,6 @@ export default {
   components: {
     Title,
     ProgressBar,
-    PopupStep,
     HintStep1
   },
   computed: {
@@ -395,6 +392,8 @@ export default {
     grid-column: 1/2;
     grid-row: 2/3;
     overflow: auto;
+    margin: -15px;
+    padding: 15px;
     .itemDesc {
       margin: 0 0 15px;
       font-size: 20px;
