+++ pages/tutorial/step6.vue
@@ -4,19 +4,41 @@
       <Title />
     </div>
 
-    <div class="wrapStep4">
+    <div class="wrapStep6">
       <div class="itemDesc">
-        Выбирите, какие из особенностей есть в вашем объекте.
+        В строке напишите адрес расположения строения, где находится ваше жильё.
+        Номер квартиры указывать не нужно.
       </div>
-
-      <div class="checkboxGrid">
-        <Checkbox
-          v-for="(opt, index) in roomOptions"
-          :key="index"
-          :desc="opt.sRoomOptionTitle"
-        />
+      <div class="wrapLocation">
+        <b-form-group
+          id="locationLabel"
+          class="form-group-restork"
+          label="Местоположение"
+          label-for="locationInput"
+        >
+          <b-form-input
+            id="locationInput"
+            placeholder=""
+            type="text"
+          ></b-form-input>
+        </b-form-group>
       </div>
-      <pre>{{ roomOptions }}</pre>
+      <div id="map">
+        <client-only>
+          <yandex-map
+            :coords="object.aObjectCoordinate.coordinates"
+            :controls="[]"
+            :zoom="14"
+            :draggable="true"
+          >
+            <ymap-marker
+              :coords="object.aObjectCoordinate.coordinates"
+              marker-id="123"
+              hint-content="Расположение объекта"
+            />
+          </yandex-map>
+        </client-only>
+      </div>
     </div>
     <div class="wrapHint">
       <HintStep1
@@ -45,23 +67,23 @@
     <div class="wrapProgress">
       <ProgressBar />
     </div>
-    <PopupStep />
   </div>
 </template>
 
 <script>
 import ProgressBar from '~/components/Tutorial/ProgressBar'
 import Title from '~/components/Tutorial/Title'
-import PopupStep from '~/components/Tutorial/PopupStep'
 import HintStep1 from '~/components/Tutorial/HintStep1'
-import Checkbox from '~/components/Tutorial/Checkbox'
 
 export default {
   data() {
     return {
-      currentStepNumber: 1,
-      iObjectTypeID: 0,
-      iCategoryID: 0,
+      object: {
+        aObjectCoordinate: {
+          type: 'Point',
+          coordinates: [43.6617826, 40.315748]
+        }
+      },
       roomCategory: [
         { id: 1, title: 'Single' },
         { id: 2, title: 'Double' },
@@ -74,18 +96,7 @@ export default {
   components: {
     Title,
     ProgressBar,
-    PopupStep,
-    HintStep1,
-    Checkbox
-  },
-  computed: {
-    roomOptions() {
-      return this.$store.state.roomOption.list
-    }
-  },
-  methods: {},
-  async asyncData({ store }) {
-    await store.dispatch('roomOption/GET_LIST')
+    HintStep1
   }
 }
 </script>
@@ -104,11 +115,13 @@ export default {
     grid-column: 1/2;
     grid-row: 1/2;
   }
-  .wrapStep4 {
+  .wrapStep6 {
     // background: red;
     grid-column: 1/2;
     grid-row: 2/3;
     overflow: auto;
+    margin: -15px;
+    padding: 15px;
     .itemDesc {
       margin: 0 0 15px;
       font-size: 20px;
@@ -122,106 +135,6 @@ export default {
         margin-top: 45px;
         @media (max-width: 767px) {
           margin-top: 37px;
-        }
-      }
-      .explanation {
-        font-size: 14px;
-        @media (max-width: 767px) {
-          display: none;
-        }
-      }
-    }
-    .singleInput {
-      width: 203px;
-      @media (max-width: 991px) {
-        width: 226px;
-      }
-      @media (max-width: 767px) {
-        width: 250px;
-      }
-    }
-    .doubleInput {
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-      grid-template-rows: 1fr;
-      grid-gap: 5px;
-      @media (max-width: 767px) {
-        grid-template-columns: 1fr 1fr;
-        grid-column-gap: 10px;
-        grid-row-gap: 20px;
-      }
-    }
-    .tripleInput {
-      display: grid;
-      grid-template-columns: 1fr 1fr 1fr;
-      grid-template-rows: 1fr;
-      grid-gap: 5px;
-      @media (max-width: 767px) {
-        grid-template-columns: 1fr 1fr;
-        grid-template-rows: 1fr 1fr;
-        grid-column-gap: 10px;
-        grid-row-gap: 20px;
-      }
-    }
-
-    .checkboxGrid {
-      margin-top: 45px;
-      display: grid;
-      grid-template-columns: 1fr 1fr;
-      grid-template-rows: 1fr;
-      grid-gap: 25px;
-      @media (max-width: 767px) {
-        grid-template-columns: 1fr;
-        grid-gap: 20px;
-      }
-    }
-    .custom-control {
-      margin: 0;
-      padding: 0;
-      background-color: white;
-      /deep/ .custom-control-input {
-        &:checked {
-          & ~ .custom-control-label {
-            opacity: 1;
-            border: 1px solid #000000;
-            &::after {
-              left: 20px;
-              background-image: url('/tutorial/tick.svg');
-              width: 25px;
-              height: 25px;
-              top: 50%;
-              transform: translate(0, -50%);
-            }
-          }
-        }
-      }
-      /deep/ .custom-control-label {
-        padding: 29px 0;
-        border: 1px solid #eaeaea;
-        border-radius: 8px;
-        width: 100%;
-        font-size: 16px;
-        line-height: 18px;
-        font-weight: 500;
-        opacity: 0.7;
-        @media (max-width: 767px) {
-          font-size: 14px;
-        }
-        &::before {
-          left: 20px;
-          border: none;
-          box-shadow: none;
-          background: none;
-          top: 50%;
-          transform: translate(0, -50%);
-        }
-        .wrapContent {
-          display: flex;
-          flex-direction: row;
-          align-items: center;
-          .image-checkbox {
-            margin: 0 30px 0 50px;
-          }
         }
       }
     }
