+++ pages/tutorial/step7.vue
@@ -6,39 +6,36 @@
 
     <div class="wrapStep7">
       <div class="itemDesc">
-        В строке напишите адрес расположения строения, где находится ваше жильё.
-        Номер квартиры указывать не нужно.
+        Перечислите какие значимые места есть рядом с вашим объектом, напишите
+        их названия и расстояния до них.
       </div>
-      <div class="wrapLocation">
-        <b-form-group
-          id="locationLabel"
-          class="form-group-restork"
-          label="Местоположение"
-          label-for="locationInput"
+
+      <PlaceInput
+        v-for="(place, index) in places"
+        :key="index"
+        :place="place"
+        :deletePlace="deletePlace"
+      />
+
+      <div class="addBtn" @click="addPlace">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="10"
+          height="10"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="#000000"
+          stroke-width="2"
+          stroke-linecap="round"
+          stroke-linejoin="round"
         >
-          <b-form-input
-            id="locationInput"
-            placeholder=""
-            type="text"
-          ></b-form-input>
-        </b-form-group>
+          <line x1="12" y1="5" x2="12" y2="19"></line>
+          <line x1="5" y1="12" x2="19" y2="12"></line>
+        </svg>
+        <span>&nbsp;&nbsp;Добавить</span>
       </div>
-      <div id="map">
-        <client-only>
-          <yandex-map
-            :coords="object.aObjectCoordinate.coordinates"
-            :controls="[]"
-            :zoom="14"
-            :draggable="true"
-          >
-            <ymap-marker
-              :coords="object.aObjectCoordinate.coordinates"
-              marker-id="123"
-              hint-content="Расположение объекта"
-            />
-          </yandex-map>
-        </client-only>
-      </div>
+
+      <pre>{{ places }}</pre>
     </div>
     <div class="wrapHint">
       <HintStep1
@@ -67,25 +64,22 @@
     <div class="wrapProgress">
       <ProgressBar />
     </div>
-    <PopupStep />
   </div>
 </template>
 
 <script>
 import ProgressBar from '~/components/Tutorial/ProgressBar'
 import Title from '~/components/Tutorial/Title'
-import PopupStep from '~/components/Tutorial/PopupStep'
 import HintStep1 from '~/components/Tutorial/HintStep1'
+import PlaceInput from '~/components/Tutorial/PlaceInput'
 
 export default {
   data() {
     return {
-      object: {
-        aObjectCoordinate: {
-          type: 'Point',
-          coordinates: [43.6617826, 40.315748]
-        }
-      },
+      places: [],
+      currentStepNumber: 1,
+      iObjectTypeID: 0,
+      iCategoryID: 0,
       roomCategory: [
         { id: 1, title: 'Single' },
         { id: 2, title: 'Double' },
@@ -98,8 +92,35 @@ export default {
   components: {
     Title,
     ProgressBar,
-    PopupStep,
-    HintStep1
+    HintStep1,
+    PlaceInput
+  },
+  computed: {
+    objectType() {
+      return this.$store.state.objectType.list
+    },
+    object() {
+      return this.$store.state.tutorial.object
+    }
+  },
+  methods: {
+    addPlace() {
+      console.log('addPlace')
+      this.places.push({})
+      // if (this.places.length) {
+      //   this.$store.dispatch(
+      //     'tutorial/ADD_PLACE',
+      //     this.places[this.places.length - 1]
+      //   )
+      // }
+    },
+    deletePlace(index) {
+      console.log('index ', index)
+      this.places.splice(index, 1)
+    }
+  },
+  async asyncData({ store, params }) {
+    await store.dispatch('objectType/GET_LIST')
   }
 }
 </script>
@@ -123,6 +144,8 @@ export default {
     grid-column: 1/2;
     grid-row: 2/3;
     overflow: auto;
+    margin: -15px;
+    padding: 15px;
     .itemDesc {
       margin: 0 0 15px;
       font-size: 20px;
@@ -136,6 +159,25 @@ export default {
         margin-top: 45px;
         @media (max-width: 767px) {
           margin-top: 37px;
+        }
+      }
+    }
+
+    .addBtn {
+      border: 1px solid #dbdbdb;
+      border-radius: 24px;
+      width: 118px;
+      height: 40px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      cursor: pointer;
+      span {
+        font-size: 12px;
+        line-height: 14px;
+        font-weight: 500;
+        @media (max-width: 767px) {
+          font-size: 11px;
         }
       }
     }
