+++ store/tutorial.js
@@ -10,6 +10,7 @@ const state = () => ({
     'Значимые места рядом',
     'Фотографии',
     'Название и описание',
+    'Статус жилья',
     'Приготовьтесь к приёму гостей'
   ],
   object: { iObjectTypeID: null },
